directly cut-paste tm code from ro
